+++ src/app/Assignment5/Question4/Q4.client.jsx
@@ -1,14 +1,14 @@
 'use client';
 
 import { useState, useEffect } from "react";
-
-export default function PaginatedList() {
+import styles from "./Q4.css";
+const  PaginatedList =()=> {
   const [data, setData] = useState([]);
   const [loading, setLoading] = useState(true);
   const [error, setError] = useState(null);
 
   const [currentPage, setCurrentPage] = useState(1);
-  const itemsPerPage = 10;
+  const itemsPerPage = 12;
 
   useEffect(() => {
     async function fetchData() {
@@ -38,63 +38,7 @@ export default function PaginatedList() {
   const handlePrev = () => setCurrentPage((p) => Math.max(p - 1, 1));
   const handleNext = () => setCurrentPage((p) => Math.min(p + 1, totalPages));
 
-  // Styles
-  const styles = {
-    container: {
-      maxWidth: "700px",
-      margin: "40px auto",
-      fontFamily: "Arial, sans-serif",
-      padding: "20px",
-      borderRadius: "8px",
-      backgroundColor: "#fefefe",
-      boxShadow: "0 2px 10px rgba(0,0,0,0.1)",
-    },
-    heading: {
-      textAlign: "center",
-      marginBottom: "20px",
-      color: "#222",
-    },
-    list: {
-      listStyleType: "none",
-      paddingLeft: 0,
-      marginBottom: "20px",
-    },
-    listItem: {
-      padding: "12px",
-      borderBottom: "1px solid #ddd",
-      fontSize: "16px",
-      color: "#333",
-    },
-    pagination: {
-      display: "flex",
-      justifyContent: "center",
-      gap: "10px",
-    },
-    button: {
-      padding: "8px 16px",
-      fontSize: "16px",
-      borderRadius: "5px",
-      border: "none",
-      backgroundColor: "#0070f3",
-      color: "white",
-      cursor: "pointer",
-      userSelect: "none",
-    },
-    buttonDisabled: {
-      backgroundColor: "#999",
-      cursor: "not-allowed",
-    },
-    error: {
-      color: "red",
-      textAlign: "center",
-      fontWeight: "bold",
-    },
-    loading: {
-      textAlign: "center",
-      fontSize: "18px",
-      fontWeight: "600",
-    },
-  };
+  
 
   return (
     <div style={styles.container}>
@@ -146,3 +90,4 @@ export default function PaginatedList() {
     </div>
   );
 }
+export default PaginatedList
