+++ src/app/Assignment5/Question7/Q7.css.jsx
@@ -1,0 +1,62 @@
+const styles = {
+    container: {
+      maxWidth: '500px',
+      margin: '40px auto',
+      padding: '20px',
+      borderRadius: '8px',
+      boxShadow: '0 4px 10px rgba(0,0,0,0.1)',
+      backgroundColor: '#fff',
+      fontFamily: "'Segoe UI', Tahoma, Geneva, Verdana, sans-serif",
+    },
+    heading: {
+      textAlign: 'center',
+      marginBottom: '20px',
+      color: '#222',
+    },
+    formGroup: {
+      marginBottom: '15px',
+      display: 'flex',
+      flexDirection: 'column',
+    },
+    label: {
+      marginBottom: '6px',
+      fontWeight: '600',
+      fontSize: '16px',
+      color: '#333',
+    },
+    input: {
+      padding: '10px',
+      fontSize: '16px',
+      borderRadius: '5px',
+      border: '1px solid #ccc',
+      outline: 'none',
+      transition: 'border-color 0.3s',
+    },
+    textarea: {
+      padding: '10px',
+      fontSize: '16px',
+      borderRadius: '5px',
+      border: '1px solid #ccc',
+      outline: 'none',
+      minHeight: '100px',
+      resize: 'vertical',
+      transition: 'border-color 0.3s',
+    },
+    button: {
+      width: '100%',
+      padding: '12px',
+      fontSize: '18px',
+      fontWeight: '700',
+      borderRadius: '6px',
+      border: 'none',
+      backgroundColor: '#0070f3',
+      color: 'white',
+      cursor: 'pointer',
+      transition: 'background-color 0.3s',
+    },
+    buttonHover: {
+      backgroundColor: '#005bb5',
+    }
+  };
+
+  export default styles
